Copy over revisions from the trunk:
  ------------------------------------------------------------------------
  r48601 | cmaynard | 2013-03-27 14:41:56 -0700 (Wed, 27 Mar 2013) | 3 lines
  Changed paths:
     M /trunk/asn1/gsm_map/packet-gsm_map-template.c
     M /trunk/epan/dissectors/packet-gsm_map.c

  Fix another incorrectly duplicated display filter, this time for hf_gsm_map_etsi_pdp_type_number.
  #BACKPORT(1.6,1.8)
  ------------------------------------------------------------------------
  r48616 | cmaynard | 2013-03-28 07:56:58 -0700 (Thu, 28 Mar 2013) | 3 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-kdsp.c

  Fix the incorrectly duplicated display filter for hf_kdsp_source_name.
  #BACKPORT(1.8)
  ------------------------------------------------------------------------
  r48619 | cmaynard | 2013-03-28 08:07:32 -0700 (Thu, 28 Mar 2013) | 3 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-sasp.c

  first_flag should be static.
  #BACKPORT(1.6,1.8)
  ------------------------------------------------------------------------
  r48620 | cmaynard | 2013-03-28 08:28:12 -0700 (Thu, 28 Mar 2013) | 7 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-ua3g.c

  Fix Coverity CID's 281433 and 281434: Unintended sign extension. (18 instances)
  Fix Coverity CID 702381: Missing break in switch.

  Add an "XXX" note regarding the framing_rtp assignment, since something appears to be missing.

  #BACKPORT(1.8)
  ------------------------------------------------------------------------
  r48622 | cmaynard | 2013-03-28 09:02:41 -0700 (Thu, 28 Mar 2013) | 3 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-bat.c

  Fix Coverity CID's 280305-280306: Negative loop bound.  (These are yet more annoying instances of not properly dealing with tvb_reported_length_remaining() possibly returning -1.)
  #BACKPORT(1.6,1.8)
  ------------------------------------------------------------------------


Copy over with manual intervention:

  ------------------------------------------------------------------------
  r48602 | cmaynard | 2013-03-27 15:04:46 -0700 (Wed, 27 Mar 2013) | 3 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-ixveriwave.c

  Fix the incorrectly duplicated display filter for hf_ixveriwave_vw_info_retryCount.
  #BACKPORT(1.8)
  ------------------------------------------------------------------------
  r48613 | cmaynard | 2013-03-28 07:23:07 -0700 (Thu, 28 Mar 2013) | 5 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-gsm_a_common.c

  Properly dissect lat and long according to 3GPP TS 23.032.  Fix incorrect display filters for hf_gsm_a_geo_loc_deg_of_lat and hf_gsm_a_geo_loc_deg_of_long.

  Fixes https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8532
  #BACKPORT(1.6,1.8)
  ------------------------------------------------------------------------


Update the release notes.

svn path=/trunk-1.8/; revision=49236

+++ epan/dissectors/packet-ua3g.c
@@ -810,8 +810,9 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
 
                             if (parameter_length <= 8) {
                                 guint64 param_value = 0;
+
                                 for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append(strbuf, val_to_str_const((guint8)(param_value), str_compressor, "Default Codec"));
                             } else {
@@ -832,7 +833,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append(strbuf, STR_ON_OFF((guint8)(param_value)));
                             } else {
@@ -850,7 +851,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 /* XXX:  Only 0x01 ==> "Yes" ??  */
                                 ep_strbuf_append(strbuf, ((guint8)param_value == 0x01) ? "Yes": "No");
@@ -874,7 +875,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append(strbuf, ((guint8)param_value) ? "Send DTMF" : "Don't Send DTMF");
                             } else {
@@ -892,7 +893,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append(strbuf, (((guint8)param_value) == 0x00) ? "Enable" : "Disable"); /* XXX: OK ? */
                             } else {
@@ -910,7 +911,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append(strbuf, (((guint8)param_value) == 0x10) ? "Enable" : "Disable");
                             } else {
@@ -948,7 +949,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                             } else {
@@ -1023,9 +1024,8 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                             if (parameter_length <= 8) {
                                 guint64 param_value = 0;
 
-                                for (i = parameter_length; i > 0; i--)
-                                {
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                for (i = parameter_length; i > 0; i--) {
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                             } else {
@@ -1104,7 +1104,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                             } else {
@@ -1301,7 +1301,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                             } else {
@@ -1508,7 +1508,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                         {
                             if ((parameter_length > 0) && (parameter_length <= 8)) {
                                 for (i = parameter_length; i > 0; i--) {
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                             } else if (parameter_length > 8) {
@@ -1580,7 +1580,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                         guint64 param_value = 0;
 
                         for (i = parameter_length; i > 0; i--) {
-                            param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                            param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                         }
                         ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                     } else {
@@ -1679,7 +1679,7 @@ decode_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb, packet_info *pinfo
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {
-                                    param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                             } else {
@@ -3729,7 +3729,7 @@ decode_cs_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb,
                                     guint64 param_value = 0;
 
                                     for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                        param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                        param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                     }
                                     ep_strbuf_append(strbuf,
                                         val_to_str_const((guint8)(param_value), str_compressor, "Default Codec"));
@@ -3783,7 +3783,7 @@ decode_cs_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb,
                                     guint64 param_value = 0;
 
                                     for (i = parameter_length; i > 0; i--) {
-                                        param_value += (tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
+                                        param_value += ((guint64)tvb_get_guint8(tvb, offset + 1 + i) << (8 * (parameter_length - i)));
                                     }
                                     ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                                 } else {
@@ -3965,7 +3965,7 @@ decode_cs_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb,
                                     guint64 param_value = 0;
 
                                     for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                        param_value += (tvb_get_guint8(tvb, offset + 2 + i) << (8 * (parameter_length - i)));
+                                        param_value += ((guint64)tvb_get_guint8(tvb, offset + 2 + i) << (8 * (parameter_length - i)));
                                     }
                                     ep_strbuf_append(strbuf,
                                         val_to_str_const((guint8)(param_value), str_compressor, "Default Codec"));
@@ -3998,7 +3998,7 @@ decode_cs_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb,
                                     guint64 param_value = 0;
 
                                     for (i = parameter_length; i > 0; i--) {  /* XXX: Not needed since only LO byte used ? */
-                                        param_value += (tvb_get_guint8(tvb, offset + 2 + i) << (8 * (parameter_length - i)));
+                                        param_value += ((guint64)tvb_get_guint8(tvb, offset + 2 + i) << (8 * (parameter_length - i)));
                                     }
                                     ep_strbuf_append(strbuf,
                                         val_to_str_const((guint8)(param_value), str_voice_mode, "Unknown"));
@@ -4145,7 +4145,9 @@ decode_cs_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb,
                             }
                         case 0x14: /* Received Framing (ms) */
                             {
+                                /* XXX: What is the point of this assignment? */
                                 framing_rtp = tvb_get_guint8(tvb, offset + 3);
+                                /* FALLTHROUGH */
                             }
                         case 0x01: /* Date Of End Of Communication */
                         case 0x02: /* Node Number */
@@ -4171,7 +4173,7 @@ decode_cs_ip_device_routing(proto_tree *tree _U_, tvbuff_t *tvb,
                                 guint64 param_value = 0;
 
                                 for (i = parameter_length; i > 0; i--) {
-                                    param_value += (tvb_get_guint8(tvb, offset + 2 + i) << (8 * (parameter_length - i)));
+                                    param_value += ((guint64)tvb_get_guint8(tvb, offset + 2 + i) << (8 * (parameter_length - i)));
                                 }
                                 ep_strbuf_append_printf(strbuf, "%" G_GINT64_MODIFIER "u", param_value);
                                 break;
